backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
base58==2.1.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Deprecated==1.2.18
ECPy==1.2.5
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.3
semantic-version==2.10.0
setuptools-rust==1.11.1
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
types-Deprecated==1.2.15.20250304
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
xrpl-py==4.1.0
